Prépare le déploiement du monorepo chez Scalingo:
- packaging de l'app react pour être embarqué dans l'api
- le chemin racine de l'api sert le bundle js (app react)
- configure un environnement docker pour le développement
