+++ mon-aide-cyber-api/.eslintrc.yml
@@ -4,6 +4,7 @@ env:
 extends:
   - eslint:recommended
   - plugin:@typescript-eslint/recommended
+overrides: []
 parser: "@typescript-eslint/parser"
 parserOptions:
   ecmaVersion: latest
