+++ mon-aide-cyber-api/package.json
@@ -1,25 +1,28 @@
 {
   "name": "mon-aide-cyber-api",
   "version": "0.0.0",
-  "description": "API de Mon Aide Cyber",
-  "type": "module",
+  "description": "API de MonAideCyber",
   "scripts": {
     "build": "tsc",
-    "start": "node --require dotenv/config index.js",
+    "start": "ts-node --transpile-only --require dotenv/config index.ts",
+    "dev": "nodemon --require dotenv/config index.ts",
     "test": "vitest run --config vitest.config.ts --reporter verbose",
     "test:coverage": "vitest --config vitest.config.ts run --coverage",
     "test:watch": "vitest --config vitest.config.ts --ui --reporter verbose"
   },
   "dependencies": {
     "dotenv": "^16.3.1",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
+    "@types/express": "^4.17.17",
+    "@typescript-eslint/eslint-plugin": "^5.60.1",
+    "@typescript-eslint/parser": "^5.60.1",
     "@vitest/coverage-v8": "^0.32.2",
     "eslint": "^8.43.0",
-    "typescript": "^5.1.3",
+    "nodemon": "^2.0.22",
     "vitest": "^0.32.2"
   }
 }
