+++ mon-aide-cyber-api/vitest.config.ts
@@ -4,6 +4,5 @@ export default defineProject({
   test: {
     environment: "node",
     globals: true,
-    reporters: ["html"],
   },
 });
